Refactor codes to reflect change in repo name i.e. from 'pos_mgmt' to 'strat_backtest'.
