+++ src/strat_backtest/base/entry_struct.py
@@ -1,7 +1,6 @@
 """Abstract class and concrete implementation of various
 entry stuctures."""
 
-import math
 from abc import ABC, abstractmethod
 from collections import deque
 from datetime import datetime
@@ -9,7 +8,7 @@ from decimal import Decimal
 
 from pydantic import ValidationError
 
-from pos_mgmt.utils import PriceAction, get_std_field
+from strat_backtest.utils import PriceAction, get_std_field
 
 from .stock_trade import StockTrade
 
@@ -233,199 +232,3 @@ class EntryStruct(ABC):
         # Validate all exit lots are less than entry lots
         if any(trade.exit_lots >= trade.entry_lots for trade in open_trades):
             raise ValueError("Completed trades observed in 'open_trades'.")
-
-
-class MultiEntry(EntryStruct):
-    """Allows multiple positions of same ticker i.e. new 'StockTrade' pydantic
-    objects with same entry lots.
-
-    - Add new long/short positions even if existing long/short positions.
-    - Number of lots entered are fixed to 'self.num_lots'.There
-
-    Usage:
-        >>> open_trades = deque()
-        >>> ticker = "AAPL"
-        >>> dt = datetime(2025, 4, 11)
-        >>> entry_price = 200.0
-        >>> ent_sig = "buy"
-        >>> multi_entry = MultiEntry(num_lots=1)
-        >>> open_trades = multi_entry.open_new_pos(
-                open_trades, ticker, dt, entry_price, ent_sig
-            )
-    """
-
-    def open_new_pos(
-        self,
-        open_trades: deque[StockTrade],
-        ticker: str,
-        dt: datetime | str,
-        ent_sig: PriceAction,
-        entry_price: float,
-    ):
-        """Generate new 'StockTrade' object populating 'ticker', 'entry_date',
-        'entry_lots' and 'entry_price'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            ticker (str):
-                Stock ticker to be traded.
-            dt (datetime | str):
-                Trade datetime object or string in "YYYY-MM-DD" format.
-            ent_sig (PriceAction):
-                Entry signal i.e. "buy", "sell" or "wait" to create new position.
-            entry_price (float):
-                Entry price for stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-        """
-
-        # Create StockTrade object to record new long/short position
-        # based on 'ent_sig'
-        if not (
-            stock_trade := self._create_new(
-                open_trades, ticker, dt, ent_sig, entry_price
-            )
-        ):
-            return open_trades
-
-        open_trades.append(stock_trade)
-        self._validate_open_trades(open_trades)
-
-        return open_trades
-
-
-class MultiHalfEntry(EntryStruct):
-    """Allows multiple positions of same ticker i.e. new 'StockTrade' pydantic
-    objects with entry lots decreasing by half with each multiple entry.
-
-    - Add new long/short positions even if existing long/short positions.
-    - Number of lots entered decreases from 'self.num_lots' by half e.g.
-    10 -> 5 -> 2 -> 1
-    - Only if 'self.num_lots' > 1; and no fractional lots allowed.
-
-    Usage:
-        >>> open_trades = deque()
-        >>> ticker = "AAPL"
-        >>> dt = datetime(2025, 4, 11)
-        >>> entry_price = 200.0
-        >>> ent_sig = "buy"
-        >>> multi_entry = MultiHalfEntry(num_lots=1)
-        >>> open_trades = multi_entry.open_new_pos(
-                open_trades, ticker, dt, entry_price, ent_sig
-            )
-    """
-
-    def open_new_pos(
-        self,
-        open_trades: deque[StockTrade],
-        ticker: str,
-        dt: datetime | str,
-        ent_sig: PriceAction,
-        entry_price: float,
-    ) -> deque[StockTrade]:
-        """Generate new 'StockTrade' object populating 'ticker', 'entry_date',
-        'entry_lots' and 'entry_price'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            ticker (str):
-                Stock ticker to be traded.
-            dt (datetime | str):
-                Trade datetime object or string in "YYYY-MM-DD" format.
-            ent_sig (PriceAction):
-                Entry signal i.e. "buy", "sell" or "wait" to create new position.
-            entry_price (float):
-                Entry price for stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-        """
-
-        # Get number of lots to enter for new position
-        entry_lots = self.get_half_lots(open_trades)
-
-        # Create StockTrade object to record new long/short position
-        # based on 'ent_sig'
-        if stock_trade := self._create_new(
-            open_trades, ticker, dt, ent_sig, entry_price, entry_lots
-        ):
-            open_trades.append(stock_trade)
-            self._validate_open_trades(open_trades)
-
-        return open_trades
-
-    def get_half_lots(self, open_trades: deque[StockTrade]) -> Decimal:
-        """Get half of latest StockTrade object in 'open_trades' till
-        minimum 1 lot."""
-
-        return (
-            math.ceil((open_trades[-1].entry_lots - open_trades[-1].exit_lots) / 2)
-            if len(open_trades) > 0
-            else self.num_lots
-        )
-
-
-class SingleEntry(EntryStruct):
-    """Allows only 1 position of same ticker i.e. no new open position created
-    if there is existing open position.
-
-    - Number of lots entered are fixed to 'self.num_lots'.
-
-    Usage:
-        >>> open_trades = deque()
-        >>> ticker = "AAPL"
-        >>> dt = datetime(2025, 4, 11)
-        >>> entry_price = 200.0
-        >>> ent_sig = "buy"
-        >>> single_entry = SingleEntry(num_lots=1)
-        >>> open_trades = single_entry.open_new_pos(
-                open_trades, ticker, dt, entry_price, ent_sig
-            )
-    """
-
-    def open_new_pos(
-        self,
-        open_trades: deque[StockTrade],
-        ticker: str,
-        dt: datetime | str,
-        ent_sig: PriceAction,
-        entry_price: float,
-    ):
-        """Generate new 'StockTrade' object populating 'ticker', 'entry_date',
-        'entry_lots' and 'entry_price'.
-
-        Args:
-            open_trades (deque[StockTrade]):
-                Deque list of StockTrade pydantic object to record open trades.
-            ticker (str):
-                Stock ticker to be traded.
-            dt (datetime | str):
-                Trade datetime object or string in "YYYY-MM-DD" format.
-            ent_sig (PriceAction):
-                Entry signal i.e. "buy", "sell" or "wait" to create new position.
-            entry_price (float):
-                Entry price for stock ticker.
-
-        Returns:
-            open_trades (deque[StockTrade]):
-                Updated deque list of 'StockTrade' objects.
-        """
-
-        if len(open_trades) > 0:
-            # No new position added since there is existing position
-            return open_trades
-
-        # Create StockTrade object to record new long/short position
-        # based on 'ent_sig'
-        if stock_trade := self._create_new(
-            open_trades, ticker, dt, ent_sig, entry_price
-        ):
-            open_trades.append(stock_trade)
-            self._validate_open_trades(open_trades)
-
-        return open_trades
